requirements.txt
anyio==4.3.0
argcomplete==2.0.0
arrow==1.3.0
attrs==23.1.0
autoflake==2.3.0
bcrypt==4.0.1
Beaker==1.12.1
beautifulsoup4==4.12.3
binaryornot==0.4.4
black==24.2.0
blivet==3.8.2
blivet-gui==2.5.0
Brlapi==0.8.5
build==1.0.3
cached-property==1.5.2
cappa==0.16.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
cookiecutter==2.5.0
cruft==2.15.0
cryptography==41.0.7
cssbeautifier==1.15.1
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
distlib==0.3.8
distro==1.8.0
djlint==1.34.1
dnf==4.18.2
docker==6.1.3
docker-compose==1.29.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
editables==0.5
EditorConfig==0.12.4
falco-cli==0.12.2
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.13.1
fros==1.1
gitdb==4.0.11
GitPython==3.1.42
h11==0.14.0
hatch==1.9.3
hatch-pip-compile==1.10.1
hatchling==1.21.1
honcho==1.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
httpcore==1.0.3
httpx==0.26.0
humanize==3.13.1
hyperlink==21.0.0
idna==3.4
isort==5.13.2
jaraco.classes==3.3.0
jeepney==0.8.0
Jinja2==3.1.3
jsbeautifier==1.15.1
json5==0.9.17
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
keyring==23.11.0
langtable==0.0.64
libcomps==0.1.20
libdnf==0.72.0
lxml==4.9.3
Mako==1.2.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
more-itertools==8.13.0
mypy-extensions==1.0.0
nftables==0.1
olefile==0.46
packaging==23.1
paho-mqtt==2.0.0
paramiko==3.4.0
parso==0.8.3
Paste==3.5.3
pathspec==0.12.1
pexpect==4.8.0
pid==2.2.3
pillow==10.2.0
pip-tools==7.4.0
platformdirs==4.2.0
pluggy==1.2.0
ply==3.11
productmd==1.38
ptyprocess==0.7.0
pwquality==1.4.5
pyasn1==0.5.1
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.45.2
pyenchant==3.2.2
pyflakes==3.2.0
Pygments==2.17.2
PyGObject==3.46.0
pykickstart==3.48
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyparted==3.13.0
pyproject_hooks==1.0.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-environ==0.4.54
python-meh==0.51
python-slugify==8.0.4
pyudev==0.24.1
pyxdg==0.27
PyYAML==6.0.1
referencing==0.30.2
regex==2023.12.25
requests==2.28.2
requests-file==1.5.1
requests-ftp==0.3.1
rich==13.7.0
rpds-py==0.17.1
rpm==4.19.1.1
SecretStorage==3.3.3
selinux @ file:///builddir/build/BUILD/libselinux-3.5/src
sepolicy @ file:///builddir/build/BUILD/selinux-3.5/python/sepolicy
setools==4.4.3
setuptools==67.7.2
shellingham==1.5.4
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sos==4.6.0
soupsieve==2.5
systemd-python==235
Tempita==0.5.2
text-unidecode==1.3
texttable==1.6.4
tomli_w==1.0.0
tomlkit==0.12.3
tqdm==4.66.2
trove-classifiers==2023.10.18
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==1.26.18
userpath==1.9.1
virtualenv==20.25.0
websocket-client==1.3.3
wheel==0.42.0
zstandard==0.21.0
